Add Active Directory SSO documentation
This moves the security documentation to a folder and adds setup instructions for Single Sign On. This covers Azure Active Directory in detail, and Azure B2C to a surface level. I will expand on B2C as the module matures.

It also adds general configuration information for SSO support, and basic information on setting up setting Drupal roles in Azure Active Directory. I'll build on this as we progress the current project where this is a requirement with any gotchas.