Update geo-technical.md (#254)
Some content design for clarification. 

We need a lot more about which Geocoder to use on prod. 
We seem to have a suggestion to use Places API (free for councils??) , but the example given is ArcGIS. This is a cognitive disconnect (in other words, I'm confused).

An "In the wild" sites list would be useful.